Extract 2tc cache methods into helper
